requirements.txt
streamlit
pdfplumber
openai
python-dotenv
python-docx